requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.3
pandas==0.25.2
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
six==1.12.0
Werkzeug==0.16.0
requests==2.22.0
